refactor: ThemeMode 를 domain/model 로 이동, UI 의존 분리
storage / data / 도메인 레이어가 ui/theme 패키지를 import 하던 어색함 정리.

- domain/model/ThemeMode 신규 (순수 enum, labelResId 제거)
- ui/theme/Theme.kt 의 ThemeMode 정의 제거하고 domain 측에서 import
- 유일한 labelResId 사용처 (SettingsPage) 는 그 자리에서 직접 when 분기로 stringResource 호출 (1회용 매핑이라 별도 확장 추출은 premature abstraction)
- 영향 받는 13개 파일의 import 일괄 갱신 (storage / data / feature / test)
