+++ app/src/main/java/com/wafflestudio/snutt2/data/user/UserRepository.kt
@@ -3,8 +3,8 @@ package com.wafflestudio.snutt2.data.user
 import com.wafflestudio.snutt2.data.Result
 import com.wafflestudio.snutt2.domain.model.PushPreferences
 import com.wafflestudio.snutt2.domain.model.SocialProviders
+import com.wafflestudio.snutt2.domain.model.ThemeMode
 import com.wafflestudio.snutt2.domain.model.User
-import com.wafflestudio.snutt2.ui.theme.ThemeMode
 import kotlinx.coroutines.flow.StateFlow
 
 interface UserRepository {
