+++ app/src/main/java/com/wafflestudio/snutt2/domain/model/ThemeMode.kt
@@ -1,0 +1,7 @@
+package com.wafflestudio.snutt2.domain.model
+
+enum class ThemeMode {
+    DARK,
+    LIGHT,
+    AUTO,
+}
