+++ app/src/main/java/com/wafflestudio/snutt2/feature/settings/ServiceInfoPage.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.viewinterop.AndroidView
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.wafflestudio.snutt2.R
+import com.wafflestudio.snutt2.domain.model.ThemeMode
 import com.wafflestudio.snutt2.ui.components.compose.SimpleTopBar
-import com.wafflestudio.snutt2.ui.theme.ThemeMode
 
 @Composable
 fun ServiceInfoPage(
