+++ app/src/main/java/com/wafflestudio/snutt2/ui/theme/Theme.kt
@@ -2,7 +2,6 @@ package com.wafflestudio.snutt2.ui.theme
 
 import android.content.Context
 import android.content.res.Configuration
-import androidx.annotation.StringRes
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.Colors
 import androidx.compose.material.MaterialTheme
@@ -11,7 +10,7 @@ import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.ui.graphics.Color
-import com.wafflestudio.snutt2.R
+import com.wafflestudio.snutt2.domain.model.ThemeMode
 
 private val LightThemeColors @Composable get() = lightColors(
     primary = SNUTTColors.White,
@@ -41,12 +40,6 @@ private val DarkThemeColors @Composable get() = darkColors(
 
 val Colors.onSurfaceVariant: Color
     get() = if (isLight) SNUTTColors.DarkerGray else SNUTTColors.Gray30
-
-enum class ThemeMode(@param:StringRes val labelResId: Int) {
-    DARK(R.string.theme_mode_dark),
-    LIGHT(R.string.theme_mode_light),
-    AUTO(R.string.theme_mode_auto),
-}
 
 val LocalThemeState = compositionLocalOf { ThemeMode.AUTO }
