+++ app/src/test/java/com/wafflestudio/snutt2/fake/FakeUserRepository.kt
@@ -4,8 +4,8 @@ import com.wafflestudio.snutt2.data.Result
 import com.wafflestudio.snutt2.data.user.UserRepository
 import com.wafflestudio.snutt2.domain.model.PushPreferences
 import com.wafflestudio.snutt2.domain.model.SocialProviders
+import com.wafflestudio.snutt2.domain.model.ThemeMode
 import com.wafflestudio.snutt2.domain.model.User
-import com.wafflestudio.snutt2.ui.theme.ThemeMode
 import kotlinx.coroutines.flow.MutableStateFlow
 
 class FakeUserRepository : UserRepository {
